requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-background-tasks>=1.2.5
django-crispy-forms==1.11.2